refactor(utils): share boolean env parsing
